refactor:restructure code, added comments, init REQWEST_CLIENT
