+++ .env.example
@@ -1,4 +1,4 @@
-# Rename this file to .env for development.toml
+# Rename this file to .env for development
 PGHOST=localhost
 PGPORT=5432
 PGUSER=musicbrainz
@@ -7,4 +7,4 @@ PGDATABASE=musicbrainz_db
 DATABASE_URL=postgres://musicbrainz:musicbrainz@localhost:5432/musicbrainz_db
 INTERNET_ARCHIVE_ACCESS_KEY=your-internet-archive-access-key
 INTERNET_ARCHIVE_SECRET=your-internet-archive-secret
-RUN_MOD=development
+RUN_MODE=development
