+++ Cargo.toml
@@ -20,6 +20,7 @@ clap = {version = "4.5.7", features = ["color", "derive"]}
 colorize = "0.1.0"
 thiserror = "1.0.61"
 config = "0.14.0"
+once_cell = "1.19.0"
 
 [lib]
 path = "src/lib.rs"
