+++ src/archival/mod.rs
@@ -1,3 +1,6 @@
+pub mod archival_response;
+mod client;
+pub mod error;
 pub mod listener;
 pub mod notifier;
 pub mod utils;
