+++ src/archival/utils.rs
@@ -1,9 +1,9 @@
-use crate::configuration::Settings;
-use crate::structs::archival_network_response::{ArchivalHtmlResponse, ArchivalResponse};
-use crate::structs::error::ArchivalError;
+use crate::archival::archival_response::{ArchivalHtmlResponse, ArchivalResponse};
+use crate::archival::client::REQWEST_CLIENT;
+use crate::archival::error::ArchivalError;
 use crate::structs::internet_archive_urls::InternetArchiveUrls;
-use reqwest::{header, Client};
 use sqlx::PgPool;
+use std::sync::Arc;
 
 ///This function is used to find the row in internet_archive_urls from where we can start the archival task
 /// The notify function will start picking URLs from the returned row id
@@ -78,41 +78,18 @@ pub async fn make_archival_network_request(
     url: &str,
     endpoint_url: &str,
 ) -> Result<ArchivalResponse, ArchivalError> {
-    let settings = Settings::new().expect("Config settings are not configured properly");
-    let mut headers = header::HeaderMap::new();
-    headers.insert("Accept", "application/json".parse().unwrap());
-    headers.insert(
-        "Authorization",
-        format!(
-            "LOW {}:{}",
-            settings.wayback_machine_api.myaccesskey, settings.wayback_machine_api.mysecret
-        )
-        .parse()
-        .unwrap(),
-    );
-    headers.insert(
-        "Content-Type",
-        "application/x-www-form-urlencoded".parse().unwrap(),
-    );
-
-    let client = Client::builder()
-        .redirect(reqwest::redirect::Policy::none())
-        .build()
-        .unwrap();
-
+    let client = Arc::clone(&REQWEST_CLIENT);
     let response = client
         .post(endpoint_url)
-        .headers(headers)
         .body(format!("url={}", url))
         .send()
         .await?;
-
-    // let response_status = response.status();
     let response_text = response.text().await?;
-
+    // Success response, contains job_id
     if let Ok(res) = serde_json::from_str::<ArchivalResponse>(&response_text) {
         return Ok(res);
     }
+    // HTML response, case when IA can not archive URL, and is under maintenance
     Ok(ArchivalResponse::Html(ArchivalHtmlResponse {
         html: response_text,
     }))
