+++ src/structs/mod.rs
@@ -1,3 +1,1 @@
-pub mod archival_network_response;
-pub mod error;
 pub mod internet_archive_urls;
